use cached, previous state
Being new to React, I am not sure of the exact behavior of setState, but based on the docs it seems to be an asynchronous function:

"Accessing this.state after calling this method can potentially return the existing value."

it seems that this.state.username could potentially be the old username value, or the new username value ('') after setting the state on line 16.  Therefore, it seems safer to use our cached value.  Is that an accurate statement to make? Am I understanding setState correctly?

Great lessons btw. Very clear and concise!